cilium: remove rules when encryption is disabled
[ upstream commit 07c857a34aac082f99c65a16b226b9561d58660f ]

Rules need to be correctly deleted otherwise its possible if in
the future cilium agent picks up a new IP address and then
re-enables encryption we may have some conflicting entries in
the route table.

Also its just cleaner to have only correct entires in the table
attempting to debug a system with multiple invalid routes is
more difficult than it needs to be.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>