cilium: encrypt iface without fib lookup is not usable
We can not do a redirect to the egress interface if the fib lookup
is not supported. If we do try to do the redirect there is no guarentee
the MAC addresses are correct because we didn't rewrite them from the
BPF datapath. This may work, if the network is very fogiving, but more
likely the NIC will drop the packet because of MAC addresses being
incorrect.

Instead if FIB lookup is not supported (<4.17 kernel) use stack context.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>