cilium: encryption, use cilium_host IP for IPSec header srcIP
When using subnet mode for encryption IPs subnets are not bound to
any specific node. So rather than use xfrm stack to write src/dst
IPs in the IPSec path we use bpf_host.o attached to cilium_host
qdisc to complete the header. This is necessary because we want
to avoid having an entry per pod in the xfrm tables and instead
have at max an entry per node.

Currently, we use the srcIP of the first interface we find in
the egress interface list. This works on some more forgiving
networks, but EKS uses source based IP routing, but does not
have a source IP in the source routing tables for all network
facing interfaces. What happens then is we rewrite the SIP of
the IPSec header to the IPAddr assigned to the network interface.
This is passed to the stack for routing. The routing table
doesn't have a source entry for the SIP so it uses the default
route. The default may or may not point to the first interface.
If it doesn't then the network appears to do a firewall rule
where the NIC/SIP lookup is unexpected and the network drops
the packet either by firewall or just missing routing rules.
It appears common for the first interface "eth0" to match the
default route so often get working by depending on this
ordering, buts its very fragile and easy to break.

Fix by using the SIP of cilium_host which always has an
entry in the source routing table.

Note: Above only applies to non-fib case (kernels <4.17) with
fib lookup we do a lookup in BPF datapath and write MACs
and IPs correctly without stack intervention.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>